+++ zego/incremental.go
@@ -5,8 +5,8 @@ import (
 )
 
 type IncrementalTicketArray struct {
-	Count         int      `json:"count"`
-	EndTime       uint     `json:"end_time"`
+	Count         int64    `json:"count"`
+	EndTime       uint64   `json:"end_time"`
 	Created       string   `json:"created"`
 	Next_page     string   `json:"next_page"`
 	Previous_page string   `json:"previous_page"`
