+++ zego/macros.go
@@ -1,19 +1,17 @@
 package zego
 
-
 type Macro struct {
-    Id           int `json:"id"`
-    Title        string `json:"title"`
-    Active       bool `json:"active"`
-    Actions      interface{} 
-    Restriction  Macro_Restriction `json:"restriction"`
+	Id          int64  `json:"id"`
+	Title       string `json:"title"`
+	Active      bool   `json:"active"`
+	Actions     interface{}
+	Restriction Macro_Restriction `json:"restriction"`
 }
 
 type Macro_Restriction struct {
-    Type         string `json:"type"`
-    Id           int `json:"id"`
+	Type string `json:"type"`
+	Id   int64  `json:"id"`
 }
-
 
 func (a Auth) ListMacros() (*Resource, error) {
